Correct line-endings on replies.
Replies were being sent using puts, which does two writes: One with
the reply, and one with a "\n".  This is wrong in two ways.  First,
the line ending is supposed to be "\r\n".  Second, doing two writes
invokes the Nagle algorithm, which hurts latency.  By sending the
reply with the correct line ending in one write, there's no need to
disable the Nagle algorithm.